refactor: Finish integrating vcpu into rec
VCPU and REC (Realm Execution Context) have almost the same purpose.
The commit aims to keep only REC which is defined by spec.

Related to #230

[Summary of changes]

- Use `vcpu_index` to record vcpuid
- Move current() to rec.rs
- Move remove() to registry.rs
- Rename reset_vcpu() as reset_ctx()
- Rename create_vcpu() as prepare_args()
