+++ rmm/src/realm/mod.rs
@@ -4,4 +4,3 @@ pub mod mm;
 pub mod rd;
 pub mod registry;
 pub mod timer;
-pub mod vcpu;
